[refactor] #32 - pages 패키지 정리
