+++ src/pages/cardViews/CardManagementView.tsx
