+++ src/router/RouterList.tsx
@@ -5,31 +5,31 @@
  */
 
 import { ReactElement } from 'react';
-import CalendarView from "@/pages/CalendarView";
-import SearchIdView from "@/pages/SearchIdView";
-import YourIdView from "@/pages/YourIdView";
-import ResetPwView from "@/pages/ResetPwView";
-import NewPwView from "@/pages/NewPwView";
-import CardRecommendView from "@/pages/CardRecommendView";
-import CardManagementView from "@/pages/CardManagementView";
-import AddCardView from "@/pages/AddCardView";
-import CardAddCompleteView from "@/pages/CardAddCompleteView";
-import DiaryView from "@/pages/DiaryView";
-import DiaryEmotionView from "@/pages/DiaryEmotionView";
-import DiaryWriteView from "@/pages/DiaryWriteView";
-import GoalSetupView from "@/pages/goal/GoalSetupView";
-import GoalEditView from "@/pages/goal/GoalEditView";
-import AchievementHistoryView from "@/pages/goal/AchievementHistoryView";
-import AchievementDetailView from "@/pages/goal/AchievementDetailView";
-import MyPageView from "@/pages/MyPageView";
-import UserInfoView from "@/pages/UserInfoView";
-import LoginView from "@/pages/LoginView";
-import SignUpAgreementView from "@/pages/SignUpAgreementView";
-import SignUpView from "@/pages/SignUpView";
-import SignUpSuccessView from "@/pages/SignUpSuccessView";
-import SignUpFailView from "@/pages/SignUpFailView";
-import CardRecomView from "@/pages/CardRecomViews/CardRecomView";
-import HomeView from "@/pages/HomeView";
+import CalendarView from "@/pages/calendarViews/CalendarView";
+import SearchIdView from "@/pages/accountViews/SearchIdView";
+import YourIdView from "@/pages/accountViews/YourIdView";
+import ResetPwView from "@/pages/accountViews/ResetPwView";
+import NewPwView from "@/pages/accountViews/NewPwView";
+import CardRecommendView from "@/pages/cardViews/CardOverView";
+import CardManagementView from "@/pages/cardViews/CardManagementView";
+import AddCardView from "@/pages/cardViews/AddCardView";
+import CardAddCompleteView from "@/pages/cardViews/CardAddCompleteView";
+import DiaryView from "@/pages/calendarViews/DiaryView";
+import DiaryEmotionView from "@/pages/calendarViews/DiaryEmotionView";
+import DiaryWriteView from "@/pages/calendarViews/DiaryWriteView";
+import GoalSetupView from "@/pages/goalViews/GoalSetupView";
+import GoalEditView from "@/pages/goalViews/GoalEditView";
+import AchievementHistoryView from "@/pages/goalViews/AchievementHistoryView";
+import AchievementDetailView from "@/pages/goalViews/AchievementDetailView";
+import MyPageView from "@/pages/myPageViews/MyPageView";
+import UserInfoView from "@/pages/myPageViews/UserInfoView";
+import LoginView from "@/pages/accountViews/LoginView";
+import SignUpAgreementView from "@/pages/accountViews/signUp/SignUpAgreementView";
+import SignUpView from "@/pages/accountViews/signUp/SignUpView";
+import SignUpSuccessView from "@/pages/accountViews/signUp/SignUpSuccessView";
+import SignUpFailView from "@/pages/accountViews/signUp/SignUpFailView";
+import CardRecomView from "@/pages/cardRecomViews/CardRecomView";
+import HomeView from "@/pages/homeViews/HomeView";
 import ReportView from "@/pages/reportViews/reportView";
 import NotificationView from "@/pages/NotificationView";
 import CategoryTop5View from "@/pages/CategoryTop5View";
